Re-organize files
https://www.youtube.com/watch?v=e9uTV44oOiw&list=PLlrATfBNZ98dC-V-N3m0Go4deliWHPFwT&index=49&pp=iAQB
