+++ Hana/src/Hana/Core/Input.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Hana/Core.h"
+#include "Hana/Core/Core.h"
 
 namespace Hana
 {
