+++ Hana/src/Hana/Events/Event.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "hnpch.h"
-#include "Hana/Core.h"
+#include "Hana/Core/Core.h"
 
 namespace Hana
 {
