+++ Hana/src/Hana/ImGui/ImGuiLayer.cpp
@@ -4,7 +4,7 @@
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_opengl3.h"
 
-#include "Hana/Application.h"
+#include "Hana/Core/Application.h"
 
 // TEMPORARY
 #include "GLFW/glfw3.h"
