+++ Hana/src/Platform/Windows/WindowsInput.cpp
@@ -1,7 +1,7 @@
 #include "hnpch.h"
 #include "WindowsInput.h"
 
-#include "Hana/Application.h"
+#include "Hana/Core/Application.h"
 #include <GLFW/glfw3.h>
 
 namespace Hana
