+++ Sandbox/imgui.ini
@@ -31,7 +31,7 @@ Size=93,48
 Collapsed=0
 
 [Window][Settings]
-ViewportPos=667,544
+ViewportPos=72,268
 ViewportId=0x1C33C293
 Size=337,188
 Collapsed=0
