+++ scripts/Win-GenProjects.bat
@@ -1,5 +1,5 @@
 @echo off
+pushd ..\
 call vendor\bin\premake\premake5.exe vs2022
-IF %ERRORLEVEL% NEQ 0 (
-    PAUSE
-)
+popd
+PAUSE
